stResolveEnemyDice implemented successfully! Not only can you roll dice with stRollEnemyDice(), but now we can resolve the dice with the stResolveEnemyDice()! We need some more work on the resolve enemy attack functions but the main skeleton is in place.
It gets the current attack dice data from the database, interprets the
results with 'dieMappings' in the material file, sorts the results into
the proper order, and resolves them one by one. Each die resolution is a
function call that is designed to handle that die roll. If the same
result occurs multiple times, each one will be a distinct function call.
To add a new enemy to the game, it needs its info filled out in the
material file, and a function for each attack in the Game.php file
(there's a section at the bottom for these) and they must be named
resolve{Enemy}Attack1 and resolve{Enemy}Attack2 and there should also be
a react to damage function.

Good progress! Keep up the good work and don't lose hope!