Several small bug fixes :) Fixed several small bugs Ive been noticing lately and added a few clarifying comments to the code. Fairly stable release.
Bugs:
-Fixed the SQL syntax error in stResolveBucketHelper that I was stuck on!
Just missing a ` so easy fix once I noticed that
-Game won't start with any enemy besides Kraken: I was missing entries
in the enemyInfo table for other enemies. Simply added that and we
should be good to go!

Other:
Added a few comments to material.inc.php and modules\php\Game.php to
clarify intent (Specifically explaining the enemyInfo fields and some of the logic
inside stResolveEnemyDice)