Bucket is done! My plan of breaking up the actions into their subactions worked great! That gives the perfect amount of control to enforce the order actions occur in and duplicity. I am globals FLAG and COUNTER to control the flow within the player's turn.
FLAG indicates current stage (true for first part, false for second)
COUNTER indicates duplicity (2 would mean 2 draws or 2 discards for
example)

The game state transitions are also applied in the perfect time, in an
if clause at the end of the appropriate action (an if at the end of
discard changes the state out of STATE_RESOLVE_BUCKET)

All in all a huge success! Should also make it straightforward to
implement the other 3 actions since there is partial overlap! :D