Started implementing Fire action Fire action partial implementation, quite a bit more to do. I focused on the FSM, arg, and actFire actions so far. $this->fireCannons is partially implemented: It assigns a random number to each cannon. Still need to interpret the results and act on them, as well as implement shoot ye treasure and pass.
I am wondering how to assciate each die with the corresponding cannon.
Posssible options include:
-a separate field location_arg that is a foreign key to cannons table
-manually assigning the die_id to be a foreign key to cannons table
-a totally different location_arg that gives its index in the column?

I'm liking the sound of the second option, but the first could do too.
If we do go with a foreign key to cannons table, the location field may
be unnecessary as the die location is totally tied to the cannon
location (both its location and location_arg, together a perfect
identifier without any other info)