Fire partially done Working enough to determine which cannons are operable and roll their corresponding dice.
I decided to manually assign the die_id of cannon dice to perfectly match
the card_id corresponding cannon (so a die_id of 5 is the die for cannon
with card_id 5, and the strengths match).

This gives us several benefits. Its a clear bijection between cannons
and their dice, so why not match that structure in the database? That
means that the die location is not stored in the dice table but in the
cannons table, meaning when you move the cannon thats also moving the
correct die. This also makes for some clean code and sql queries focused
on "dice inner join cannon".

Next up is actually damaging the enemy!