Blank slate Getting ready for attempting to migrate stocks to the bga-cards...
The current stocks code has been commented out and I began with the
setup described on the bga documentation page on bga-cards. Just giving
us a clean slate to work with

I hope that using the new official javascript class for displaying cards
will make it easier to work with, update, and animate. It does have
built in animations for all of the common events like moving and
rotating cards.

I am investigating the game French Tarot as an example of using
bga-cards as well as its demo page. I am confused about several things
but the main one is how the source image is linked in. I think it must
be linked to the card manager? If not how would all of the stocks under
its control know what the cards look like? I'm just confused tho because
it isnt totally clear.