bga-cards starting to make sense I finally found what I couldn't understand about bga-cards: how did they indicate what image file to use and how to display the correct card image?
The image src is in the CSS! The card stock adds a class to the
elements, and they added that background to that class in the css. The
card manager is told how to identify the proper image by defining
functions that give position based on card info.

Started implementing rudimentary water column... as in they are solid
blue and static cards. And not overlapping...

Next up: figuring out how to overlap them. I think I need manual
position stock but I am having trouble defining it... Progress city!