Fine tuned spacing! Fixed the spacing between busted cannons and breaches, + teeny spacings
I attempted to set up the spacing between the busted cannons and
breaches in a fancy way. I set the onCardCountChange property of the
busted cannons stock to adjust the margin top of the breach div. Seems
to work but I wonder how smooth the animations will be? Might need to
revisit this later but works for now.

Made minor adjustments to the spacing of different elements,
mostly by teeny changes in margin padding of the decks to account for
the deck's shadow.