I was soooo close! Dice color works great now! I figured out the critical error: The function that actually attaches the data-color tag to the divs.. it does not do that by default but instead you set up a function to do so when you instantiate the dice manager.
Another good note: It turns out that apparent visual artifact, the one I
almost reported as a bug with bga-cards where the die shows a little bit
of other faces on the edges, making it look sloppy: its a feature not a
bug! And really a technically impressive feature that shows some of the
3Dness of a die by showing a bit of the side as well, exaggerated more
the further from the center the die is located and adjustable with the
perspective variable.