Removed dial icon, working on actDraw and it notif Commented out the dial icon code for now, can put it back if we decide to later but im not getting much feedback either way at the moment. I'll probably revisit that later in alpha testing.
Working on improving actDraw and notif_actDraw, particularly focusing on
getting the notifications right and setting up the notifaction handler
to modify the front end correctly. The goal is to display the correct
information to all users (current player and other players), not reveal
more than we must to frontend, and setup the updates so that you don't
have to keep hitting refresh.

We are making definite progress but we have not yet achieved the third
goal of eliminating refreshes... Especially for the discard notifs
(which are currently missing). It now displays the item names/obfuscates
the item names appropriately in the notifications (when someone draws a
facedown card, it tells that player what they got and tells everyone
else they drew an unknown card. When someone draws a faceup card, it
tells everyone exactly what they took). Progress in the notification
handler, but not quite there yet. It feels really close though...