Very happy with draw/discard actions/notifs! Found and fixed the bugs!!! Mercilessly squashed them! No worries or struggles ever again... maybe. Well see what the future has in store for us.
Whitespace above new cards:
It was the whitespace that ManualPositionStock put there to fan out the
cards! We simply had to remove the whitespace that ManualPositionStock
added when cards. Ended up just adding the property onCardRemoved that
resets the whitespace when cards are removed from the manual position
stocks!

Could not remove cards from the manual position stock to update everyone
else display (not the active player):
Simply had a variable wrong. I think it was switching on notif.source
and it needed to be notif.card.location. Simply changed the structure of
the notif in the backend and failed to update the frontend
accordingly...

Bug 1 + bug 2 fixed = New one popped up!

Now when attempting to remove cards on everyone elses screen, came up
with a null error!:
in fix #1 we were correcting the style of the removed card, but since it
was successfully removed from fix #2, that card no longer existed on the
other clients. Had to add a check for null!

Now the drawing and discarding cards works great! The discard pile looks
good and is upside down, cards move around and flip as needed, clients
update properly from notifs and dont need refreshed every time. The
recent bug fix in correctGapUnderBoard feels great now! Doing great!
Keep up the good work!