Fixed the cannon bug! Identified and fixed the cannon bug: rolling a cannon when there were no cannons left caused the front end to crash: there were no checks or safeguards on the backend side so it proceeded as usual. When it failed to find a card, it simply returned with a nonsensical id of 0, but the frontend was woefully unprepared. It would then attempt to move the cannon with id 0 to the breaches column. Upon finding no such cannon exists, it would crash with an runtime error.
The fix: Simple and reliable if to check if there are any cards in the
cannon column on the backend. If there were none, it notifies the
frontend there were no cannons to damage and stops there (to make it
abundantly clear to the players what happened: they still get a notif,
it just reports why nothing happened). If it does find a cannon, then
proceed as usual. I did leave the check in the helper function. In
theory two checks is redundant, but who knows what the future holds, or
if there will ever be a way to damage cannons besides dice. I also
updated the nonsensical return value in the helper function to be more
obviously nonsence (-1) to make it clear to future me if I ever face
bugs on cannons

Simple bug: simple solution: Always add the safety checks. Even if you
think future you is smart enough to give it proper input. And make the
error messages and safety handling thorough and clear.