First couple card actions + working on card order Added a bit to the itemsDoThingsIGuess function and attempted to fix some of my card bugs. I think I eliminated one of the sources of confusion, but it would seem there are more lurking in the deep...
Added some safety checks and two items: bottleORum and boneClub.
Cards:
The symptoms: Cards do not occur in the proper order, their order is not
even super stable. Now this may not be a huge problem for the water
column thats all random anyways, but the treasure and cannon
columns need to be ordered for rules enforcement (queue structure).
Upon inspection, the obvious source of the problem is that they didn't
have an proper order in the database either! Their location args were
nonsensical, mostly 0s.

I eliminated several similar functions that moved cards around
(addToCannonColumn, addToTreasureColumn, pickForWaterColumn,
removeFromCannonColumn) and simplified to addToColumn and
removeFromColumn. The theory is that I can eliminate inconsistencies and
confusions by refactoring to use only these two instead of many
variations. If the issues persist, I need only change it in one clear
place, not all over the place. Obviously the next step was to replace
all occurrences of the old function calls with these new ones. Then
there were some less obvious places where I used the deck functions
instead of mine to move cards around (which also contriubuted to the
problem!) and converted those to addToColumn as well.

The result: The cards operation is quite a bit simpler and cleaner, at
least on backend. It is much more obvious and straightforward to
deal/remove cards. They have a much better structure in database! It
looks more like 4,2,0,-2,-4,etc. instead of 0,0,1,0,0,...! Improvement.
Possibly I want to adjust the addToColumn to make these positive
ascending or something but we can decide later if that is necessary.

New problems: They still aren't rendered in the proper order on the
front end! At least it seems stable. I think javascript properties are
not ordered, so I will need to communicate in some other way the
order... maybe an explicit field (card_order array?).

I am also concerned because it would appear the removeFromColumn doesnt
do anything, (possibly removeFromCannonsColumn, its predecessor, hasn't
for a while?? I'm confused tho, it looks like it works?)