Found and squashed the column order bug! Turns out it was multifaceted and was lurking in several unexpected places... but no matter! We shall prevail! The good guys always win... right? The columns display cards in their proper order. Using the addToColumn successfully adds the card in the appropriate order (the bottom of the column).
The first half of the bug was the previous commit: the database
had no idea what the order was supposed to be! The location_arg was a
disaster, mostly zero or nonsensical values! After some reorganizing, I
unified the backend to use addToColumn to add cards, instead of several
different approaches that yielded mixed results.

The second half of the bug was more subtle: poor communication between
front and backend. Specifically, some simple changes in the getAllDatas
rectified the issue. It all came down to this: The backend wasn't
sending the cards in the proper order, and the order is stored slightly
differently in PHP vs JS. On php side, the cards were an array with ids
as keys and cards as values. On javascript side, these cards were
received as objects with fields matching the ids, each pointing to an
object of the card data. Now the technicality is that the php array was
ordered but javascript objects don't preserve/guarantee order for their
members. To avoid this, javascript needed to receive simple arrays
(which are ordered), which amounted to the backend needing a
array_values and only sending cards. For some unknown reason, backend
also needs to use array_reverse to get the order right, mabye its sort
is backwards? But it works now!