Several bug fixes! And a bit of Kraken progress! Found and fixed a few bugs. Worked on Kraken's special die notifications and code a bit.
Bug #1: Game crashing bug in actDraw that prevented a draw action from
completing sucessfully. Somehow even though the signature looked right
in the setCardOrientation and the call definitely sent a false, the
setCardOrientatin recieved null for $faceUp instead of true. Still not
entirely sure why? Maybe an oddity of php? Fixed with a simple conditon
checking for true value and fixing it if it isnt right.

Bug #2: SQL query for updating card_face_up failed because it wanted an
int not a bool? So I had to give it either a 1 or 0 depending on the
truth of the $faceUp...

Bug #3: The animation for adding a new breach resulted in a facedown
breach, (which fixed itself when refreshed indicating it was a notif
issue not a state issue). Fixed by sending the whole card not just the
id in the notif, because it uses the location field to decide whether it
is faceup or down.

Bug #4: Kraken's splash attack resulted in a clear water being displayed
to the user. Upon taking it, they would actually get the treasure it
really is. Another faceUp or down bug, just needed to add a
setCardOrientation to fix

Additionally, filled out the Krakens attack code and gave both special
attacks notifs! Next up is it might need a notif handler for
resolveKrakenAttack1 and actually implementing the angered attack in
full.