Major bug fixed! No refreshing to progress action The bug: Every time there is an action with multiple parts, you need to refresh between parts because the game doesn't automatically update and tell you what needs done next. Draw 2 discard 2 turns into draw 1, refresh, draw 1, refresh, discard 1, refresh, discard 1...
This has been an annoying bug that has eluded me for some time. There
was recently a relevant discussion in the developer discord that gave me
the insight to what exactly was happening here: When an action was
complete, it moved the game state to the next state. When an action was
incomplete (still needed a draw, discard, etc...) it simply updated
the FLAG/COUNTER to indicate current game state. But that means it
didn't update the front end with the new possible moves and
parameters... so we needed to update the frontend somehow. The best
option was to add a self-referencing state change to the FSM called
'again'. Updating the gamestate to again results in the same state you
were already in, but it updates the frontend of the current
state/options/parameters. The alternative was a notification that told
the front the updated information but that seems unnecessarily clunky
and awkward.